Fix flutter-elinux snap build failure
This commit fixes an issue with the flutter build of the
workshops application. The override-build scriptlet for
the part terminates prematurely, however snapcraft doesn't
indicate that the part's build stage failed. Moving the
first flutter-elinux (which supports many of the same
commands as flutter) command to an override-stage part
resolves the problem.

Note - the flutter-src part doesn't actually result in a
fully working toolchain. An actually build command triggers
the toolchain to be downloaded/built.

Signed-off-by: Tony Espy <espy@canonical.com>